tests/krate/publish: Extract `rate_limit` module
